+++ soil/simulation.py
@@ -15,7 +15,7 @@ from nxsim import NetworkSimulation
 
 from . import serialization, utils, basestring, agents
 from .environment import Environment
-from .serialization import logger
+from .utils import logger
 from .exporters import for_sim as exporters_for_sim
 
 
@@ -65,8 +65,6 @@ class Simulation(NetworkSimulation):
         whose value indicates the state
     dir_path: str, optional
         Directory path to load simulation assets (files, modules...)
-    outdir : str, optional
-        Directory path to save simulation results
     seed : str, optional
         Seed to use for the random generator
     num_trials : int, optional
@@ -87,11 +85,11 @@ class Simulation(NetworkSimulation):
 
     def __init__(self, name=None, group=None, topology=None, network_params=None,
                  network_agents=None, agent_type=None, states=None,
-                 default_state=None, interval=1, dump=None, dry_run=False,
-                 outdir=None, num_trials=1, max_time=100,
-                 load_module=None, seed=None, dir_path=None,
-                 environment_agents=None, environment_params=None,
-                 environment_class=None, **kwargs):
+                 default_state=None, interval=1, num_trials=1,
+                 max_time=100, load_module=None, seed=None,
+                 dir_path=None, environment_agents=None,
+                 environment_params=None, environment_class=None,
+                 **kwargs):
 
         self.seed = str(seed) or str(time.time())
         self.load_module = load_module
@@ -101,18 +99,10 @@ class Simulation(NetworkSimulation):
         self.num_trials = num_trials
         self.max_time = max_time
         self.default_state = default_state or {}
-        if not outdir:
-            outdir = os.path.join(os.getcwd(),
-                                       'soil_output')
-        self.outdir = os.path.join(outdir,
-                                   self.group or '',
-                                   self.name)
         self.dir_path = dir_path or os.getcwd()
         self.interval = interval
-        self.dump = dump
-        self.dry_run = dry_run
 
-        sys.path += list(x for x in [self.outdir, os.getcwd(), self.dir_path] if x not in sys.path)
+        sys.path += list(x for x in [os.getcwd(), self.dir_path] if x not in sys.path)
 
         if topology is None:
             topology = serialization.load_network(network_params,
@@ -142,6 +132,7 @@ class Simulation(NetworkSimulation):
         return self.run(*args, **kwargs)
 
     def run(self, *args, **kwargs):
+        '''Run the simulation and return the list of resulting environments'''
         return list(self._run_simulation_gen(*args, **kwargs))
 
     def _run_sync_or_async(self, parallel=False, *args, **kwargs):
@@ -152,7 +143,7 @@ class Simulation(NetworkSimulation):
                            **kwargs)
             for i in p.imap_unordered(func, range(self.num_trials)):
                 if isinstance(i, Exception):
-                    logger.error('Trial failed:\n\t{}'.format(i.message))
+                    logger.error('Trial failed:\n\t%s', i.message)
                     continue
                 yield i
         else:
@@ -162,29 +153,30 @@ class Simulation(NetworkSimulation):
                                      **kwargs)
 
     def _run_simulation_gen(self, *args, parallel=False, dry_run=False,
-                            exporters=None, **kwargs):
+                            exporters=None, outdir=None, exporter_params={}, **kwargs):
+        logger.info('Using exporters: %s', exporters or [])
+        logger.info('Output directory: %s', outdir)
         exporters = exporters_for_sim(self,
-                                      exporters or [])
+                                      exporters or [],
+                                      dry_run=dry_run,
+                                      outdir=outdir,
+                                      **exporter_params)
+
         with utils.timer('simulation {}'.format(self.name)):
-            if not (dry_run or self.dry_run):
-                logger.info('Dumping results to {}'.format(self.outdir))
-                self.dump_pickle(self.outdir)
-                self.dump_yaml(self.outdir)
-            else:
-                logger.info('NOT dumping results')
             for exporter in exporters:
                 exporter.start()
 
             for env in self._run_sync_or_async(*args, parallel=parallel,
-                                               dry_run=dry_run, **kwargs):
+                                               **kwargs):
                 for exporter in exporters:
-                    exporter.env(env)
+                    exporter.trial_end(env)
                 yield env
 
             for exporter in exporters:
                 exporter.end()
 
     def get_env(self, trial_id = 0, **kwargs):
+        '''Create an environment for a trial of the simulation'''
         opts = self.environment_params.copy()
         env_name = '{}_trial_{}'.format(self.name, trial_id)
         opts.update({
@@ -192,19 +184,17 @@ class Simulation(NetworkSimulation):
             'topology': self.topology.copy(),
             'seed': self.seed+env_name,
             'initial_time': 0,
-            'dry_run': self.dry_run,
             'interval': self.interval,
             'network_agents': self.network_agents,
             'states': self.states,
             'default_state': self.default_state,
             'environment_agents': self.environment_agents,
-            'outdir': self.outdir,
         })
         opts.update(kwargs)
         env = self.environment_class(**opts)
         return env
 
-    def run_trial(self, trial_id=0, until=None, dry_run=False, **opts):
+    def run_trial(self, trial_id=0, until=None, **opts):
         """Run a single trial of the simulation
 
         Parameters
@@ -214,13 +204,9 @@ class Simulation(NetworkSimulation):
         # Set-up trial environment and graph
         until = until or self.max_time
         env = self.get_env(trial_id = trial_id, **opts)
-        dry_run = self.dry_run or dry_run
         # Set up agents on nodes
         with utils.timer('Simulation {} trial {}'.format(self.name, trial_id)):
             env.run(until)
-        if self.dump and not dry_run:
-            with utils.timer('Dumping simulation {} trial {}'.format(self.name, trial_id)):
-                env.dump(formats = self.dump)
         return env
     def run_trial_exceptions(self, *args, **kwargs):
         '''
@@ -240,24 +226,25 @@ class Simulation(NetworkSimulation):
     def to_yaml(self):
         return yaml.dump(self.to_dict())
 
-    def dump_yaml(self, outdir = None, file_name = None):
-        outdir = outdir or self.outdir
-        if not os.path.exists(outdir):
-            os.makedirs(outdir)
-        if not file_name:
-            file_name=os.path.join(outdir,
-                                     '{}.dumped.yml'.format(self.name))
-        with open(file_name, 'w') as f:
+
+    def dump_yaml(self, f=None, outdir=None):
+        if not f and not outdir:
+            raise ValueError('specify a file or an output directory')
+
+        if not f:
+            f = os.path.join(outdir, '{}.dumped.yml'.format(self.name))
+
+        with utils.open_or_reuse(f, 'w') as f:
             f.write(self.to_yaml())
 
-    def dump_pickle(self, outdir = None, pickle_name = None):
-        outdir = outdir or self.outdir
-        if not os.path.exists(outdir):
-            os.makedirs(outdir)
-        if not pickle_name:
-            pickle_name=os.path.join(outdir,
-                                       '{}.simulation.pickle'.format(self.name))
-        with open(pickle_name, 'wb') as f:
+    def dump_pickle(self, f=None, outdir=None):
+        if not outdir and not f:
+            raise ValueError('specify a file or an output directory')
+
+        if not f:
+            f = os.path.join(outdir,
+                             '{}.simulation.pickle'.format(self.name))
+        with utils.open_or_reuse(f, 'wb') as f:
             pickle.dump(self, f)
 
     def __getstate__(self):
@@ -279,8 +266,6 @@ class Simulation(NetworkSimulation):
     def __setstate__(self, state):
         self.__dict__ = state
         self.load_module = getattr(self, 'load_module', None)
-        if self.outdir not in sys.path:
-            sys.path += [self.outdir, os.getcwd()]
         if self.dir_path not in sys.path:
             sys.path += [self.dir_path, os.getcwd()]
         self.topology = json_graph.node_link_graph(state['topology'])
@@ -308,24 +293,14 @@ def from_config(conf_or_path):
     return sim
 
 
-def run_from_config(*configs, outdir=None, dump=None, timestamp=False,  **kwargs):
+def run_from_config(*configs, **kwargs):
     for config_def in configs:
         # logger.info("Found {} config(s)".format(len(ls)))
         for config, path in serialization.load_config(config_def):
             name = config.get('name', 'unnamed')
             logger.info("Using config(s): {name}".format(name=name))
 
-            if timestamp:
-                sim_folder = '{}_{}'.format(name,
-                                            time.strftime("%Y-%m-%d_%H:%M:%S"))
-            else:
-                sim_folder = name
-            if dump is not None:
-                config['dump'] = dump
             dir_path = config.pop('dir_path', os.path.dirname(path))
-            outdir = config.pop('outdir', outdir)
             sim = Simulation(dir_path=dir_path,
-                             outdir=outdir,
                              **config)
-            logger.info('Dumping results to {} : {}'.format(sim.outdir, sim.dump))
             sim.run_simulation(**kwargs)